requirements.txt
streamlit>=1.30.0
pandas>=2.1.0
numpy>=1.26.0
plotly>=5.18.0
scipy>=1.12.0
statsmodels>=0.14.0
Jinja2>=3.1.0
pydantic>=2.5.0
loguru>=0.7.0
PyYAML>=6.0.0
# --- Optional (not needed for demo mode) ---
# google-cloud-bigquery>=3.14.0
# db-dtypes>=1.2.0
# pymc>=5.10.0
# arviz>=0.17.0
# WeasyPrint>=61.0
